inference/fastapi-inference/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mlflow>=3.7.0
pandas>=2.0.0
numpy<2
xgboost
pydantic>=2.0
google-cloud-storage
